pyproject.toml
[project]
name = "HostHunter"
version = "1.0.0"
description = "A host checking tool"
authors = [{name = "HanSoBoren", email = "raihanzxzy@gmail.com"}]
license = "MIT"
readme = "README.md"
requires-python = ">=3.9,<4.0"

dependencies = [
    "requests==2.32.4",
    "websocket-client==1.8.0",
    "rich==14.1.0",
    "pytest==8.4.1",
    "ruff==0.12.8",
    "black==25.1.0",
    
]

[project.scripts]
hosthunter = "src.cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
